add: register maintenance for a vehiculo

+++ src/main/java/cl/myhotel/vehicles/dto/VehiculoDTO.java
@@ -4,6 +4,14 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.*;
 
+/**
+ * Data Transfer Object (DTO) for vehicle information.
+ * This class is used to transfer vehicle-related data between layers of the application.
+ * It contains fields for the vehicle ID, brand, model, license plate,
+ * year, mileage, engine displacement, and various vehicle types.
+ *
+ * @author Leonel Zeballos
+ */
 @Getter
 @Setter
 @NoArgsConstructor
@@ -11,7 +19,7 @@ import lombok.*;
 @Builder
 @JsonInclude(Include.NON_NULL)
 public class VehiculoDTO {
-    private Integer id;
+    private Long id;
     private String marca;
     private String modelo;
     private String patente;
